Rework as an intro lecture to lab
Reworked as an interactive lecture to be done before the Tic
Tac Toe lab. Students will go through the process of creating
the project repo, planning with user stories, creating a
wireframe and scaffolding the project code together.  The
remainder of the day will be lab time to complete the game
and review possible solutions in a separate repo.